+++ Example/WalletApp/PresentationLayer/Wallet/CATransactionModal/CATransactionPresenter.swift
@@ -192,3 +192,7 @@ final class CATransactionPresenter: ObservableObject {
 
 // MARK: - SceneViewModel
 extension CATransactionPresenter: SceneViewModel {}
+
+
+
+
